+++ ExpandWorldSize/config/Patcher.cs
@@ -1,4 +1,5 @@
 using HarmonyLib;
+using Service;
 using UnityEngine;
 
 namespace ExpandWorldSize;
@@ -8,21 +9,27 @@ public static class Patcher
 {
   public static WorldGenerator? WG;
   public static bool IsMenu => WG == null || WG.m_world.m_menu;
-  private static Harmony? Harmony;
-  public static void Init(Harmony harmony)
+#nullable disable
+  private static Harmony Harmony;
+  private static Harmony DynamicHarmony;
+#nullable enable
+  public static void Init(Harmony harmony, Harmony dynamicHarmony)
   {
     Harmony = harmony;
+    DynamicHarmony = dynamicHarmony;
     Patch();
   }
   public static void Patch()
   {
-    if (Harmony == null) throw new("Harmony not initialized.");
     Harmony.UnpatchSelf();
     Harmony.PatchAll();
-    CheckWorldStretch(Harmony);
-    CheckWaterDepth(Harmony);
-    CheckForest(Harmony);
-    CheckAltitude(Harmony);
+    CheckWorldStretch(DynamicHarmony);
+    CheckWaterDepth(DynamicHarmony);
+    CheckForest(DynamicHarmony);
+    CheckAltitude(DynamicHarmony);
+    CreateAshlandsGap.Patch(DynamicHarmony, !Configuration.DeepNorthGap);
+    CreateDeepNorthGap.Patch(DynamicHarmony, Configuration.DeepNorthGap);
+    GetAshlandsHeight.Patch(DynamicHarmony, Configuration.AshlandsWidthRestriction, Configuration.AshlandsLengthRestriction);
   }
 
   [HarmonyPatch(typeof(WorldGenerator), nameof(WorldGenerator.VersionSetup)), HarmonyPostfix, HarmonyPriority(Priority.Last)]
@@ -32,45 +39,78 @@ public static class Patcher
     WorldInfo.Refresh();
   }
 
-
+  private static float PatchedWorldStretch = 1f;
   private static void CheckWorldStretch(Harmony harmony)
   {
-    if (!IsMenu && Configuration.WorldStretch != 1f) return;
+    var worldStretch = IsMenu ? 1f : WorldInfo.WorldStretch;
+    if (PatchedWorldStretch == worldStretch) return;
+    PatchedWorldStretch = worldStretch;
     var method = AccessTools.Method(typeof(WorldGenerator), nameof(WorldGenerator.GetBiomeHeight));
     var patch = AccessTools.Method(typeof(Stretch), nameof(Stretch.GetBiomeHeight));
     harmony.Unpatch(method, patch);
+    if (worldStretch != 1f)
+      harmony.Patch(method, prefix: new(patch));
     method = AccessTools.Method(typeof(WorldGenerator), nameof(WorldGenerator.GetAshlandsOceanGradient), [typeof(Vector3)]);
     patch = AccessTools.Method(typeof(Stretch), nameof(Stretch.GetAshlandsOceanGradient));
     harmony.Unpatch(method, patch);
+    if (worldStretch != 1f)
+      harmony.Patch(method, prefix: new(patch));
     method = AccessTools.Method(typeof(Minimap), nameof(Minimap.GetMaskColor));
     patch = AccessTools.Method(typeof(Stretch), nameof(Stretch.GetMaskColor));
     harmony.Unpatch(method, patch);
+    if (worldStretch != 1f)
+      harmony.Patch(method, prefix: new(patch));
     method = AccessTools.Method(typeof(WorldGenerator), nameof(WorldGenerator.GetBiome), [typeof(float), typeof(float), typeof(float), typeof(bool)]);
     patch = AccessTools.Method(typeof(Stretch), nameof(Stretch.GetBiome));
     harmony.Unpatch(method, patch);
+    if (worldStretch != 1f)
+      harmony.Patch(method, prefix: new(patch));
     method = AccessTools.Method(typeof(WorldGenerator), nameof(WorldGenerator.AddRivers));
     patch = AccessTools.Method(typeof(AddRivers), nameof(AddRivers.Prefix));
     harmony.Unpatch(method, patch);
+    if (worldStretch != 1f)
+      harmony.Patch(method, prefix: new(patch));
   }
+  private static float PatchedWaterDepthMultiplier = 1f;
+  private static float PatchedWaterLevel = 30f;
   private static void CheckWaterDepth(Harmony harmony)
   {
-    if (!IsMenu && Configuration.WaterDepthMultiplier != 1f) return;
+    var waterDepthMultiplier = IsMenu ? 1f : WorldInfo.WaterDepth;
+    if (PatchedWaterDepthMultiplier == waterDepthMultiplier && WorldInfo.WaterLevel == PatchedWaterLevel) return;
+    PatchedWaterDepthMultiplier = waterDepthMultiplier;
+    PatchedWaterLevel = WorldInfo.WaterLevel;
     var method = AccessTools.Method(typeof(WorldGenerator), nameof(WorldGenerator.GetBiomeHeight));
     var patch = AccessTools.Method(typeof(BiomeHeight), nameof(BiomeHeight.Postfix));
     harmony.Unpatch(method, patch);
+    // Water level is used in the patch but doesn't do anything without the multiplier.
+    if (waterDepthMultiplier == 1f) return;
+    harmony.Patch(method, postfix: new(patch));
   }
+  private static float PatchedForestMultiplier = 1f;
   private static void CheckForest(Harmony harmony)
   {
-    if (!IsMenu && Configuration.ForestMultiplier != 1f && Configuration.ForestMultiplier != 0f) return;
+    var forestMultiplier = IsMenu ? 1f : WorldInfo.ForestMultiplier;
+    if (PatchedForestMultiplier == forestMultiplier) return;
+    PatchedForestMultiplier = forestMultiplier;
     var method = AccessTools.Method(typeof(WorldGenerator), nameof(WorldGenerator.GetForestFactor));
     var patch = AccessTools.Method(typeof(Forest), nameof(Forest.Postfix));
     harmony.Unpatch(method, patch);
+    if (forestMultiplier == 1f) return;
+    harmony.Patch(method, postfix: new(patch));
   }
+  private static float PatchedAltitudeDelta = 0f;
+  private static float PatchedAltitudeMultiplier = 1f;
   private static void CheckAltitude(Harmony harmony)
   {
-    if (!IsMenu && (Configuration.AltitudeDelta != 0f || Configuration.AltitudeMultiplier != 1f)) return;
+    var altitudeDelta = IsMenu ? 0f : WorldInfo.BaseAltitudeDelta;
+    var altitudeMultiplier = IsMenu ? 1f : WorldInfo.AltitudeMultiplier;
+    if (PatchedAltitudeDelta == altitudeDelta && PatchedAltitudeMultiplier == altitudeMultiplier) return;
+    PatchedAltitudeDelta = altitudeDelta;
+    PatchedAltitudeMultiplier = altitudeMultiplier;
     var method = AccessTools.Method(typeof(WorldGenerator), nameof(WorldGenerator.GetBaseHeight));
     var patch = AccessTools.Method(typeof(BaseHeight), nameof(BaseHeight.Postfix));
     harmony.Unpatch(method, patch);
+    if (altitudeDelta == 0f && altitudeMultiplier == 1f) return;
+    harmony.Patch(method, postfix: new(patch));
   }
 }