+++ ExpandWorldSize/features/Poles.cs
@@ -8,30 +8,28 @@ namespace ExpandWorldSize;
 public class GetAshlandsHeight
 {
   private static readonly double DefaultWidthRestriction = 7500f;
-  private static double WidthRestriction = DefaultWidthRestriction;
+  private static double PatchedWidthRestriction = DefaultWidthRestriction;
   private static readonly double DefaultLengthRestriction = 1000f;
-  private static double LengthRestriction = DefaultLengthRestriction;
-  public static bool Patch(Harmony harmony, double widthRestriction, double lengthRestriction)
+  private static double PatchedLengthRestriction = DefaultLengthRestriction;
+  public static void Patch(Harmony harmony, double widthRestriction, double lengthRestriction)
   {
-    if (WidthRestriction == widthRestriction && LengthRestriction == lengthRestriction) return false;
+    if (PatchedWidthRestriction == widthRestriction && PatchedLengthRestriction == lengthRestriction) return;
     var method = AccessTools.Method(typeof(WorldGenerator), nameof(WorldGenerator.GetAshlandsHeight));
     var transpiler = AccessTools.Method(typeof(GetAshlandsHeight), nameof(Transpiler));
-    WidthRestriction = widthRestriction;
-    LengthRestriction = lengthRestriction;
+    PatchedWidthRestriction = widthRestriction;
+    PatchedLengthRestriction = lengthRestriction;
     harmony.Unpatch(method, transpiler);
-    if (WidthRestriction != DefaultWidthRestriction || LengthRestriction != DefaultLengthRestriction)
+    if (PatchedWidthRestriction != DefaultWidthRestriction || PatchedLengthRestriction != DefaultLengthRestriction)
       harmony.Patch(method, transpiler: new HarmonyMethod(transpiler));
-
-    return true;
   }
 
   static IEnumerable<CodeInstruction> Transpiler(IEnumerable<CodeInstruction> instructions)
   {
     return new CodeMatcher(instructions)
       .MatchForward(false, new CodeMatch(OpCodes.Ldc_R8, 1000.0))
-      .SetOperandAndAdvance(LengthRestriction)
+      .SetOperandAndAdvance(PatchedLengthRestriction)
       .MatchForward(false, new CodeMatch(OpCodes.Ldc_R8, 7500.0))
-      .SetOperandAndAdvance(WidthRestriction)
+      .SetOperandAndAdvance(PatchedWidthRestriction)
       .InstructionEnumeration();
   }
 }
@@ -39,23 +37,17 @@ public class GetAshlandsHeight
 public class CreateAshlandsGap
 {
   private static bool IsPatched = false;
-  public static bool Patch(Harmony harmony, bool doPatch)
+  public static void Patch(Harmony harmony, bool doPatch)
   {
-    if (IsPatched == doPatch) return false;
+    if (IsPatched == doPatch) return;
     var method = AccessTools.Method(typeof(WorldGenerator), nameof(WorldGenerator.CreateAshlandsGap));
     var prefix = AccessTools.Method(typeof(CreateAshlandsGap), nameof(DisableGap));
     IsPatched = doPatch;
     if (doPatch)
-    {
-      Log.Info("Patching CreateAshlandsGap");
       harmony.Patch(method, prefix: new HarmonyMethod(prefix));
-    }
     else
-    {
-      Log.Info("Unpatching CreateAshlandsGap");
       harmony.Unpatch(method, prefix);
-    }
-    return true;
+
   }
 
   static bool DisableGap(ref double __result)
@@ -68,23 +60,16 @@ public class CreateAshlandsGap
 public class CreateDeepNorthGap
 {
   private static bool IsPatched = false;
-  public static bool Patch(Harmony harmony, bool doPatch)
+  public static void Patch(Harmony harmony, bool doPatch)
   {
-    if (IsPatched == doPatch) return false;
+    if (IsPatched == doPatch) return;
     var method = AccessTools.Method(typeof(WorldGenerator), nameof(WorldGenerator.CreateDeepNorthGap));
     var prefix = AccessTools.Method(typeof(CreateAshlandsGap), nameof(DisableGap));
     IsPatched = doPatch;
     if (doPatch)
-    {
-      Log.Info("Patching CreateDeepNorthGap");
       harmony.Patch(method, prefix: new HarmonyMethod(prefix));
-    }
     else
-    {
-      Log.Info("Unpatching CreateDeepNorthGap");
       harmony.Unpatch(method, prefix);
-    }
-    return true;
   }
 
   static bool DisableGap(ref double __result)