+++ ExpandWorldSize/features/WorldSize.cs
@@ -80,6 +80,14 @@ public class SetupMaterial
   public static void Prefix(WaterVolume __instance)
   {
     __instance.m_waterSurface.material.SetFloat("_WaterEdge", Configuration.WorldTotalRadius);
+    // Zone water should be at y 30, anything above that is dungeon water.
+    if (__instance.transform.position.y < 30.001f && __instance.m_collider is BoxCollider box)
+    {
+      // Default is -20 center with 30 size, probably 10 meters extra for waves.
+      // -100 meters should give plenty of space for deeper water.
+      box.center = new Vector3(0, -100, 0);
+      box.size = new Vector3(62, 220, 64);
+    }
   }
 }